requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
PyJWT
python-multipart